Move unit test to sysvar sub-submodule
